chore: move MySQL datasource validation tests to a separate file (#27373)
